Refactor RecipeDescription and Controls components for improved type definitions and clarity; remove CoffeeTimerContent component
